fix: 重ねて当たった特典の合計を long で持つ
規則は排他ではなく重ねて当たり、1 件の点数は int の上限までしか縛られていない。
合計を int で持つと回り込んで負になり、行はすべて積まれているのに申領の応答が
「付与はありません」の分岐へ落ちる。

「1 件の仕訳は int でも、積み上がった合計は int を超えうる」は台帳が残高・受注ごとの
付与合計・巻き戻しの動いた量で既に採っている規律で、同じ概念を返す
MemberVisitResponse#grantedPoints も long である。応答の型もそちらへ揃える。

+++ backend/src/integrationTest/java/com/kizuna/point/BenefitGrantIT.java
@@ -125,7 +125,7 @@ class BenefitGrantIT extends CrossStoreTestSupport {
     assertThat(claimed.getStatusCode()).isEqualTo(HttpStatus.CREATED);
     // 応答は「この申領で記帳したポイント」。伝票の付与予定額に特典を足した額でなければ、特典だけが
     // 付いた申領が画面で「付与はありません」になる。
-    assertThat(claimed.getBody().path("granted_points").asInt())
+    assertThat(claimed.getBody().path("granted_points").asLong())
         .as("伝票の付与予定額と特典の合計を返すこと")
         .isEqualTo(EXPECTED_ORDER_GRANT + BENEFIT_POINTS);
     BenefitRow benefit = onlyBenefitGrantOf(claimant.id());
