+++ backend/src/main/java/com/kizuna/point/application/BenefitGrantService.java
@@ -43,12 +43,12 @@ public class BenefitGrantService {
    *
    * @param orderDate 根拠受注の営業日。適用期間の窓はこの日で判じる（記帳した日ではない）
    * @param actorUserId 契機を起こした主体。完了なら操作した従業員、事後申領なら申領した会員本人
-   * @return この契機で記帳した特典ポイントの合計。当たった規則が無ければ 0
+   * @return この契機で記帳した特典ポイントの合計。当たった規則が無ければ 0。1 件の規則の点数は int でも、規則は排他でなく積み上がるので合計は int を超えうる
    */
-  public int grantVisitBenefits(
+  public long grantVisitBenefits(
       long memberId, String orderId, Long storeId, LocalDate orderDate, Long actorUserId) {
     LocalDate grantedOn = LocalDate.now(ZoneId.of(appProperties.getTimezone()));
-    int granted = 0;
+    long granted = 0;
     for (BenefitRule rule : benefitRuleRepository.findByTypeAndEnabledTrue(BenefitRuleType.VISIT)) {
       if (!rule.firesFor(storeId, orderDate) || alreadyBenefited(rule, memberId, orderId)) {
         continue;
